requirements.txt
fastapi 
uvicorn
alembic
sqlalchemy
psycopg2
httpx
secrets
passlib
python-jose
secure-smtplib==0.1.1